Move ParseUrl to Util namespace
